# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dev Connect API",
    "description": "API for developer-focused social network platform where users can create profiles, showcase projects, and connect with other developers.",
    "version": "1.0.0",
    "contact": {
      "name": "Dev Connect Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "schemas": {
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "content": {
            "type": "string",
            "example": "This project looks amazing! The UI design is very clean."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorId": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "projectId": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "avatarUrl": {
                "type": "string",
                "nullable": true
              }
            }
          }
        },
        "required": [
          "id",
          "content",
          "createdAt",
          "authorId",
          "projectId",
          "author"
        ]
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "type": "object"
          }
        },
        "required": ["success", "message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "errors": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ]
          }
        },
        "required": ["success", "message"]
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Data retrieved successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "meta": {
                "type": "object",
                "properties": {
                  "page": {
                    "type": "integer",
                    "example": 1
                  },
                  "limit": {
                    "type": "integer",
                    "example": 10
                  },
                  "totalItems": {
                    "type": "integer",
                    "example": 47
                  },
                  "totalPages": {
                    "type": "integer",
                    "example": 5
                  },
                  "hasNextPage": {
                    "type": "boolean",
                    "example": true
                  },
                  "hasPrevPage": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "required": ["success", "data"]
      }
    },
    "parameters": {
...
      "projectIdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Project ID (CUID format)",
        "example": "cmas2xbuy0000tf8cp8h37lgl"
      },
      "pageParam": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 1
        },
        "description": "Page number for pagination"
      },
      "limitParam": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 10
        },
        "description": "Number of items per page"
      },
      "sortByParam": {
        "name": "sortBy",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "default": "createdAt"
        },
        "description": "Field to sort by"
      },
      "sortOrderParam": {
        "name": "sortOrder",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "enum": ["asc", "desc"],
          "default": "desc"
        },
        "description": "Sort direction (asc/desc)"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/projects/{id}/comments": {
      "get": {
        "summary": "Get project comments",
        "description": "Retrieves all comments for a specific project",
        "tags": ["Comments"],
        "parameters": [
          {
            "$ref": "#/components/parameters/projectIdParam"
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/sortByParam"
          },
          {
            "$ref": "#/components/parameters/sortOrderParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Comments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Comment"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid project ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a comment",
        "description": "Creates a new comment on a specific project",
        "tags": ["Comments"],
        "parameters": [
          {
            "$ref": "#/components/parameters/projectIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
          "type": "object",
          "properties": {
            "content": {
              "type": "string",
              "example": "This is a great project! I love the UI design."
            }
          },
          "required": ["content"]
              }
            },
            "application/json": {
              "schema": {
          "type": "object",
          "properties": {
            "content": {
              "type": "string",
              "example": "This is a great project! I love the UI design."
            }
          },
          "required": ["content"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Comment"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}